refactor: unify console command registry in core/console-commands.js
Merges tab-complete.js and the command implementations from ui/console.js into a single registry in js/core/console-commands.js. Each CommandDef now owns its verb identity, tab-completion behaviour, and execution logic in one place.

The registry is open for extension: console.js calls registerCommand() to override the "cheat" entry with browser-specific sub-commands (relayout, restore), then delegates to the core execute for everything else. tabComplete() and getCommand() read the live registry Map so overrides are picked up automatically.

Testability: getCommand(verb).execute(args) is now callable directly from tests without going through the full command-dispatch pipeline.
